feat: streamline CI checks and enhance e2e test setup for Android and iOS
