+++ checks/src/test/kotlin/com/yral/checks/AndroidE2eTest.kt
@@ -6,8 +6,11 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Order
+import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
+import java.util.concurrent.TimeUnit
 
+@Tag("e2e")
 @Order(2)
 class AndroidE2eTest {
     @BeforeEach
@@ -64,9 +67,9 @@ class AndroidE2eTest {
 
         @BeforeAll
         @JvmStatic
-        fun installApk() {
+        fun setup() {
+            startEmulator()
             // Clean install: uninstall first to wipe any retained app state, then fresh install.
-            // adb install -r keeps user data; uninstall+install guarantees a clean slate.
             exec("adb", "uninstall", appId) // ignore failure — app may not be installed yet
             execOrFail(
                 "adb", "install",
@@ -78,6 +81,50 @@ class AndroidE2eTest {
             exec("adb", "shell", "am", "start", "-n", "$appId/.MainActivity")
             Thread.sleep(3_000)
             exec("adb", "shell", "am", "force-stop", appId)
+        }
+
+        private fun startEmulator() {
+            // Wait up to 30s for any existing device (handles the case where the developer
+            // already has an emulator running before invoking the e2e task).
+            exec("adb", "reconnect")
+            val alreadyRunning = ProcessBuilder("adb", "wait-for-device")
+                .directory(repoRoot)
+                .inheritIO()
+                .start()
+                .waitFor(30, TimeUnit.SECONDS)
+            if (alreadyRunning && captureOutput("adb", "devices").lines().count { it.contains("\tdevice") } > 0) {
+                println("Android device already available — skipping emulator start.")
+                return
+            }
+
+            val avd = captureOutput("emulator", "-list-avds").trim().lines()
+                .firstOrNull { it.isNotBlank() }
+            checkNotNull(avd) { "No AVDs found — create one in Android Studio first." }
+
+            println("Starting emulator: $avd")
+            ProcessBuilder("emulator", "-avd", avd, "-no-audio", "-no-snapshot", "-no-boot-anim")
+                .directory(repoRoot)
+                .start() // background process; emulator outlives this @BeforeAll
+
+            execOrFail("adb", "wait-for-device")
+
+            val deadline = System.currentTimeMillis() + 5 * 60_000L
+            while (System.currentTimeMillis() < deadline) {
+                if (captureOutput("adb", "shell", "getprop", "sys.boot_completed").trim() == "1") {
+                    val pmDeadline = System.currentTimeMillis() + 60_000L
+                    while (System.currentTimeMillis() < pmDeadline) {
+                        if (exec("adb", "shell", "pm", "list", "packages", "android") == 0) {
+                            println("Emulator ready.")
+                            return
+                        }
+                        Thread.sleep(2_000)
+                    }
+                    println("Emulator ready (PM check timed out, proceeding anyway).")
+                    return
+                }
+                Thread.sleep(2_000)
+            }
+            error("Emulator did not finish booting within 5 minutes")
         }
     }
 }
