+++ checks/src/test/kotlin/com/yral/checks/IosE2eTest.kt
@@ -6,11 +6,13 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Order
+import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.condition.EnabledOnOs
 import org.junit.jupiter.api.condition.OS
 import java.io.File
 
+@Tag("e2e")
 @Order(3)
 @EnabledOnOs(OS.MAC)
 class IosE2eTest {
